% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busco_bar.R
\name{busco_bar}
\alias{busco_bar}
\title{Create a gg style barplot for BUSCO genes}
\usage{
busco_bar(x, max_copy = NULL)
}
\arguments{
\item{x}{A matrix of BUSCO coordinates for one chromosome/scaffold.}

\item{max_copy}{maximum copy number}
}
\value{
A ggplot object.
}
\description{
Create a gg style barplot for BUSCO genes
}
