% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCOplot-package.R
\docType{package}
\name{BUSCOplot-package}
\alias{BUSCOplot}
\alias{BUSCOplot-package}
\title{BUSCOplot: Plot Results from BUSCO}
\description{
Plot results from BUSCO (one paragraph).
}
\author{
\strong{Maintainer}: Brian J. Knaus \email{briank.lists@gmail.com} (\href{https://orcid.org/0000-0003-1665-4343}{ORCID})

}
\keyword{internal}
