% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fingerprint.R
\name{dist_fingerprint}
\alias{dist_fingerprint}
\title{Distance based on difference in BUSCO coordinates}
\usage{
dist_fingerprint(x, perc_comp = 0.2)
}
\arguments{
\item{x}{A table of BUSCO coordinates.}

\item{perc_comp}{percent composition}
}
\value{
A matrix.
}
\description{
Distance based on difference in BUSCO coordinates
}
