% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_line_map.R
\name{gg_line_map}
\alias{gg_line_map}
\title{Create a gg style line map using BUSCO coordinates}
\usage{
gg_line_map(
  busc,
  rect = FALSE,
  palpha = 1,
  size = 1,
  lalpha = 1,
  linewidth = 1,
  line_na_rm = TRUE
)
}
\arguments{
\item{busc}{A matrix of BUSCO coordinates for one chromosome/scaffold.}

\item{rect}{A number.}

\item{palpha}{A number.}

\item{size}{A number.}

\item{lalpha}{A number.}

\item{linewidth}{A number.}

\item{line_na_rm}{A number.}
}
\value{
A ggplot object.
}
\description{
Create a gg style line map using BUSCO coordinates
}
