% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_map.R
\name{line_map}
\alias{line_map}
\title{Create a line map using BUSCO coordinates}
\usage{
line_map(x, alpha = 1, ...)
}
\arguments{
\item{x}{A table of BUSCO coordinates.}

\item{alpha}{transparency or alpha channel}

\item{...}{other options.}
}
\value{
An invisible NULL.
}
\description{
Create a line map using BUSCO coordinates
}
