% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_known.R
\name{make_known}
\alias{make_known}
\title{Create list for BUSCO genes}
\usage{
make_known(x)
}
\arguments{
\item{x}{A matrix of BUSCO coordinates for one chromosome/scaffold.}
}
\value{
A list.
}
\description{
Create list for BUSCO genes
}
