% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_BUSCOS.R
\name{position_BUSCOS}
\alias{position_BUSCOS}
\title{Adjust the position and orientation of BUSCO sequences}
\usage{
position_BUSCOS(x, min_BUSCO = 0, sort_by = c("max_pos", "max_busco"))
}
\arguments{
\item{x}{A data.frame of BUSCO coordinates with Busco_ids as row.names.}

\item{min_BUSCO}{minimum number of BUSCO genes per Sequence to retain it.}

\item{sort_by}{Sort the resulting BUSCO sequences}
}
\value{
A data.frame of BUSCO coordinates with Busco_ids as row.names.
}
\description{
Adjust the position and orientation of BUSCO sequences
}
\examples{
DM6file <- system.file("extdata", "DM6_full_table.tsv.gz", package = 'BUSCOplot')
DM6 <- read_busco(DM6file)
chromL <- make_known(DM6, sample_name = "DM6:")
#lapply(chromL[1:2], head, n=2)
ATLfile <- system.file("extdata", "atlantic_full_table.tsv.gz", package = 'BUSCOplot')
ATL <- read_busco(ATLfile)
chromL <- add_unknown(known_list = chromL, unknown_df = ATL, sample_name = "ATL:")
#lapply(chromL[1:2], head, n=2)

gg_line_map(chromL[[1]])
tmp <- position_BUSCOS(chromL[[1]])
gg_line_map(tmp)

}
