% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_busco.R
\name{read_busco}
\alias{read_busco}
\title{Read in a BUSCO full_table.tsv}
\usage{
read_busco(file)
}
\arguments{
\item{file}{A full_table.tsv file created by BUSCO.}
}
\value{
A data.frame.
}
\description{
Read in a BUSCO full_table.tsv
}
