% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unknown.R
\name{add_unknown}
\alias{add_unknown}
\title{Add unknowns to list of BUSCO genes}
\usage{
add_unknown(known_list, unknown_df)
}
\arguments{
\item{known_list}{A list of known BUSCOs.}

\item{unknown_df}{A data.frame of BUSCOs}
}
\value{
A list.
}
\description{
Add unknowns to list of BUSCO genes
}
