% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_line_map.R
\name{gg_line_map}
\alias{gg_line_map}
\title{Create a gg style line map using BUSCO coordinates}
\usage{
gg_line_map(
  busc,
  rect = FALSE,
  palpha = 1,
  size = 1,
  lalpha = 1,
  linewidth = 1,
  line_na_rm = TRUE,
  check_table = TRUE
)
}
\arguments{
\item{busc}{A data.frame of BUSCO coordinates with Busco_ids as row.names.}

\item{rect}{A logical indicating if a rectangle should be drawn from the lowest to highest BUSCO.}

\item{palpha}{A number \link{0-1} for point alpha channel (transparency).}

\item{size}{A number to adjust point size.}

\item{lalpha}{A number \link{0-1} for line alpha channel (transparency).}

\item{linewidth}{A number to adjust line width.}

\item{line_na_rm}{A logical indicating if lines should end at missing points.}

\item{check_table}{Logical to perform check for a properly formatted BUSCO table.}
}
\value{
A ggplot object.
}
\description{
Create a gg style line map using BUSCO coordinates
}
\examples{
DM6file <- system.file("extdata", "DM6_full_table.tsv.gz", package = 'BUSCOplot')
DM6 <- read_busco(DM6file)
chromL <- make_known(DM6, sample_name = "DM6:")
#lapply(chromL[1:2], head, n=2)
ATLfile <- system.file("extdata", "atlantic_full_table.tsv.gz", package = 'BUSCOplot')
ATL <- read_busco(ATLfile)
chromL <- add_unknown(known_list = chromL, unknown_df = ATL, sample_name = "ATL:")
#lapply(chromL[1:2], head, n=2)

gg_line_map(chromL[[1]])


}
