% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_line_map.R
\name{gg_line_map}
\alias{gg_line_map}
\title{Create a gg style line map using BUSCO coordinates}
\usage{
gg_line_map(
  busc,
  rect = FALSE,
  palpha = 1,
  size = 1,
  lalpha = 1,
  linewidth = 1,
  line_na_rm = TRUE
)
}
\arguments{
\item{busc}{A matrix of BUSCO coordinates for one chromosome/scaffold.}

\item{rect}{A number.}

\item{palpha}{A number.}

\item{size}{A number.}

\item{lalpha}{A number.}

\item{linewidth}{A number.}

\item{line_na_rm}{A number.}
}
\value{
A ggplot object.
}
\description{
Create a gg style line map using BUSCO coordinates
}
\examples{
DM6file <- system.file("extdata", "DM6_full_table.tsv.gz", package = 'BUSCOplot')
DM6 <- read_busco(DM6file)
chromL <- make_known(DM6, sample_name = "DM6:")
#lapply(chromL[1:2], head, n=2)
ATLfile <- system.file("extdata", "atlantic_full_table.tsv.gz", package = 'BUSCOplot')
ATL <- read_busco(ATLfile)
chromL <- add_unknown(known_list = chromL, unknown_df = ATL, sample_name = "ATL:")
#lapply(chromL[1:2], head, n=2)

tmp <- chromL[[1]]
colnames(tmp)[ colnames(tmp) == "Gene_Start" ] <- unique(chromL[[1]]$Sequence)[1]
#tmp[1:3, ]
rownames(tmp) <- tmp$Busco_id
tmp <- tmp[ , c(4, 9:ncol(tmp))]
#tmp[1:3, ]

gg_line_map(tmp)


}
